Update output.md
There doesn't seem to be a `TargetType` property available, or if it does it neither auto-completes in Visual Studio nor affects a build's output type.

There IS however the property `OutputType`, which can be set to the same values as listed here under the section TargetType.

I think it's worth investigating other options listed in this documentation set, such as `OutputAssembly`. I believe this is referring to the properties `OutputPath` and `AssemblyName`. As well, the example does not indicate that a trailing slash is needed after the folder name.

The document also has "Specify the full name and extension of the file you want to create". However, it's unclear whether this simply means the base file name or the fully qualified filename. I suspect this is meant to simply be "the name and extension of the file". I think the documentation should also indicate whether or not the "folder" can be relative.

I believe there might be other documentation in this set that should be updated. For example, the [`MainEntryPoint` or `StartupObject`](https://github.com/dotnet/docs/blob/main/docs/csharp/language-reference/compiler-options/advanced.md#mainentrypoint-or-startupobject) references a `MainEntryPoint` property which does not seem to work or exist. It also incorrectly states that `StartupObject` must be a fully qualified class name.

I hope I didn't read this entire section incorrectly.

Thanks!